Moved to configs directory
