--- lua/configs/colorschemes/dracula.lua
@@ -1,4 +1,0 @@
-require 'dracula'.setup {
-  transparent_bg = true,
-  italic_comment = true
-}
